requirements.txt
streamlit
tensorflow
numpy
Pillow
langchain
langchain-community
google-serper
python-dotenv
